@@ -8,5 +8,5 @@ public class Urls {
     public static final String BASE = "https://api.themoviedb.org/";
     public static final String MOVIE_POPULAR = "/3/movie/popular";
     public static final String MOVIE_TOP_RATE = "/3/movie/top_rated";
-    public static final String IMAGE_URL = "https://image.tmdb.org/t/p/w500/";
+    public static final String IMAGE_URL = "https://image.tmdb.org/t/p/w92/";
 }
